# backend/docs/swagger.yaml
definitions:
  github_com_movie-matcher_backend_internal_models.AgeRating:
    enum:
    - g
    - pg
    - pg13
    - r
    - nc17
    - nc21
    - tvY
    - tvY7
    - tvG
    - tvPg
    - tv14
    - tvMa
    type: string
    x-enum-varnames:
    - AgeRatingG
    - AgeRatingPG
    - AgeRatingPG13
    - AgeRatingR
    - AgeRatingNC17
    - AgeRatingNC21
    - AgeRatingTVY
    - AgeRatingTVY7
    - AgeRatingTVG
    - AgeRatingTVPG
    - AgeRatingTV14
    - AgeRatingTVMA
...
  github_com_movie-matcher_backend_internal_models.AnimeSeason:
    enum:
    - winter
    - spring
    - summer
    - autumn
    type: string
    x-enum-varnames:
    - AnimeSeasonWinter
    - AnimeSeasonSpring
    - AnimeSeasonSummer
    - AnimeSeasonAutumn
  github_com_movie-matcher_backend_internal_models.AnimeSeries:
    properties:
      ageRating:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AgeRating'
      cast:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Cast'
        type: array
      comments:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeriesComment'
        type: array
      createdAt:
        type: string
      currentEpisode:
        type: integer
      description:
        type: string
      descriptionI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      episodeDuration:
        type: integer
      episodesCount:
        type: integer
      favorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeriesFavorite'
        type: array
      genres:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Genre'
        type: array
      id:
        type: integer
      images:
        items: {}
        type: array
      isHidden:
        type: boolean
      listStatus:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.ListStatus'
      lists:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeriesList'
        type: array
      poster:
        type: string
      rating:
        type: number
      ratingCount:
        type: integer
      releaseDate:
        type: string
      releaseSchedule:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.JSONMap'
      reviews:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeriesReview'
        type: array
      season:
        allOf:
        - $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeason'
        description: winter, spring, summer, autumn — сезон выхода (зима/весна/лето/осень)
      seasonNumber:
        type: integer
      similar:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeries'
        type: array
      sites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeriesSite'
        type: array
      staff:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.MediaStaff'
        type: array
      status:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.MediaStatus'
      studios:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Studio'
        type: array
      themes:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Theme'
        type: array
      title:
        type: string
      titleI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      titleKatakana:
        description: Оригинальное название на катакане (или кане)
        type: string
      titleRomaji:
        description: 'Транслитерация: Yuusha-kei ni Shosu: Choubatsu...'
        type: string
      updatedAt:
        type: string
      videos:
        items: {}
        type: array
    type: object
  github_com_movie-matcher_backend_internal_models.AnimeSeriesComment:
    properties:
      animeSeries:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeries'
      animeSeriesId:
        type: integer
      createdAt:
        type: string
      depth:
        type: integer
      id:
        type: integer
      minusCount:
        type: integer
      parent:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeriesComment'
      parentId:
        type: integer
      plusCount:
        type: integer
      replies:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeriesComment'
        type: array
      repliesCount:
        type: integer
      text:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.User'
      userId:
        type: integer
    type: object
  github_com_movie-matcher_backend_internal_models.AnimeSeriesFavorite:
    properties:
      animeSeries:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeries'
      animeSeriesId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.User'
      userId:
        type: integer
    type: object
  github_com_movie-matcher_backend_internal_models.AnimeSeriesList:
    properties:
      animeSeries:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeries'
      animeSeriesId:
        type: integer
      comment:
        type: string
      completedAt:
        type: string
      createdAt:
        type: string
      currentEpisode:
        type: integer
      currentProgress:
        type: number
      id:
        type: integer
      rating:
        type: integer
      startedAt:
        type: string
      status:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.ListStatus'
      titleReaction:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TitleReaction'
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.User'
      userId:
        type: integer
    type: object
  github_com_movie-matcher_backend_internal_models.AnimeSeriesReview:
    properties:
      actingRating:
        type: integer
      animeSeries:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeries'
      animeSeriesId:
        type: integer
      characterRating:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      musicRating:
        type: integer
      overallRating:
        type: integer
      productionRating:
        type: integer
      review:
        type: string
      reviewStatus:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.ReviewStatus'
      storyRating:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.User'
      userId:
        type: integer
      visualsRating:
        type: integer
    type: object
  github_com_movie-matcher_backend_internal_models.AnimeSeriesSite:
    properties:
      animeSeries:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeries'
      animeSeriesId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      site:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Site'
      siteId:
        type: integer
      updatedAt:
        type: string
      url:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.BookFavorite:
    properties:
      book:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Book'
      bookId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.User'
      userId:
        type: integer
    type: object
...
  github_com_movie-matcher_backend_internal_models.Cast:
    properties:
      character:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Character'
      characterId:
        type: integer
      createdAt:
        type: string
      dubbings:
        description: 'Дубляж: даббер и язык'
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Dubbing'
        type: array
      episodeNumber:
        type: integer
      id:
        type: integer
      images:
        items: {}
        type: array
      person:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Person'
      personId:
        type: integer
      poster:
        description: 'CastPoster: изображение для карточки каста'
        type: string
      role:
        type: string
      roleType:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.RoleType'
      updatedAt:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.Character:
    properties:
      avatar:
        type: string
      cast:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Cast'
        type: array
      createdAt:
        type: string
      description:
        type: string
      descriptionI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      id:
        type: integer
      images:
        items: {}
        type: array
      name:
        type: string
      nameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      updatedAt:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.Dubbing:
    properties:
      cast:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Cast'
      castId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      language:
        description: 'Код языка: ru, en, ja и т.д.'
        type: string
      person:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Person'
      personId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.Genre:
    properties:
      createdAt:
        type: string
      description:
        type: string
      descriptionI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      emoji:
        type: string
      id:
        type: integer
      mediaTypes:
        items: {}
        type: array
      name:
        type: string
      nameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      updatedAt:
        type: string
    type: object
  github_com_movie-matcher_backend_internal_models.JSONMap:
    additionalProperties: true
    type: object
...
  github_com_movie-matcher_backend_internal_models.ListStatus:
    enum:
    - watching
    - planned
    - completed
    - onHold
    - dropped
    - rewatching
    type: string
    x-enum-varnames:
    - ListStatusWatching
    - ListStatusPlanned
    - ListStatusCompleted
    - ListStatusOnHold
    - ListStatusDropped
    - ListStatusRewatching
  github_com_movie-matcher_backend_internal_models.LocalizedString:
    additionalProperties:
      type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.MediaStaff:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      mediaId:
        type: integer
      mediaType:
        type: string
      person:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Person'
      personId:
        type: integer
      profession:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Profession'
      updatedAt:
        type: string
    type: object
  github_com_movie-matcher_backend_internal_models.MediaStatus:
    enum:
    - announced
    - in_production
    - released
    - finished
    - cancelled
    - postponed
    type: string
    x-enum-comments:
      MediaStatusAnnounced: Анонсировано
      MediaStatusCancelled: Отменено
      MediaStatusFinished: Завершено
      MediaStatusInProduction: В производстве
      MediaStatusPostponed: Отложено
      MediaStatusReleased: Вышло / выходит
    x-enum-varnames:
    - MediaStatusAnnounced
    - MediaStatusInProduction
    - MediaStatusReleased
    - MediaStatusFinished
    - MediaStatusCancelled
    - MediaStatusPostponed
...
  github_com_movie-matcher_backend_internal_models.Person:
    properties:
      avatar:
        type: string
      biography:
        type: string
      biographyI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      birthDate:
        type: string
      cast:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Cast'
        type: array
      country:
        type: string
      createdAt:
        type: string
      firstName:
        type: string
      firstNameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      id:
        type: integer
      images:
        items: {}
        type: array
      lastName:
        type: string
      lastNameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      profession:
        items: {}
        type: array
      updatedAt:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.Profession:
    enum:
    - actor
    - actress
    - director
    - producer
    - writer
    - cinematographer
    - composer
    - editor
    - animator
    - voiceActor
    - author
    - illustrator
    - artist
    - gameDesigner
    - levelDesigner
    - translator
    - literaryEditor
    type: string
    x-enum-comments:
      ProfessionArtist: Художник (манга, комиксы)
      ProfessionAuthor: Книги, манга, ранобэ
      ProfessionGameDesigner: Геймдизайнер (игры)
      ProfessionIllustrator: Иллюстратор (книги, ранобэ)
      ProfessionLiteraryEditor: Литературный редактор (книги)
      ProfessionTranslator: Переводчик (книги, локализация)
    x-enum-varnames:
    - ProfessionActor
    - ProfessionActress
    - ProfessionDirector
    - ProfessionProducer
    - ProfessionWriter
    - ProfessionCinematographer
    - ProfessionComposer
    - ProfessionEditor
    - ProfessionAnimator
    - ProfessionVoiceActor
    - ProfessionAuthor
    - ProfessionIllustrator
    - ProfessionArtist
    - ProfessionGameDesigner
    - ProfessionLevelDesigner
    - ProfessionTranslator
    - ProfessionLiteraryEditor
...
  github_com_movie-matcher_backend_internal_models.ReviewStatus:
    enum:
    - neutral
    - positive
    - negative
    - surprised
    - disappointed
    - excited
    - confused
    - laugh
    - boring
    - cry
    - cool
    - cold
    - scary
    - shushing
    - exploding_head
    - face_holding_back_tears
    - kiss
    - pleading_face
    - saluting_face
    type: string
    x-enum-varnames:
    - ReviewStatusNeutral
    - ReviewStatusPositive
    - ReviewStatusNegative
    - ReviewStatusSurprised
    - ReviewStatusDisappointed
    - ReviewStatusExcited
    - ReviewStatusConfused
    - ReviewStatusLaugh
    - ReviewStatusBoring
    - ReviewStatusCry
    - ReviewStatusCool
    - ReviewStatusCold
    - ReviewStatusScary
    - ReviewStatusShushing
    - ReviewStatusExplodingHead
    - ReviewStatusFaceHoldingBackTears
    - ReviewStatusKiss
    - ReviewStatusPleadingFace
    - ReviewStatusSalutingFace
  github_com_movie-matcher_backend_internal_models.RoleType:
    enum:
    - main
    - supporting
    - cameo
    - extra
    - narrator
    - guest
    - background
    type: string
    x-enum-varnames:
    - RoleTypeMain
    - RoleTypeSupporting
    - RoleTypeCameo
    - RoleTypeExtra
    - RoleTypeNarrator
    - RoleTypeGuest
    - RoleTypeBackground
  github_com_movie-matcher_backend_internal_models.Site:
    properties:
      createdAt:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
  github_com_movie-matcher_backend_internal_models.SocialLinksMap:
    additionalProperties:
      type: string
    type: object
  github_com_movie-matcher_backend_internal_models.Studio:
    properties:
      country:
        type: string
      createdAt:
        type: string
      description:
        type: string
      descriptionI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      id:
        type: integer
      name:
        type: string
      nameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      poster:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.Theme:
    properties:
      createdAt:
        type: string
      description:
        type: string
      descriptionI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      emoji:
        type: string
      id:
        type: integer
      mediaTypes:
        items: {}
        type: array
      name:
        type: string
      nameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      updatedAt:
        type: string
    type: object
  github_com_movie-matcher_backend_internal_models.TitleReaction:
    enum:
    - surprised
    - disappointed
    - sad
    - joyful
    - inspiring
    - scary
    - funny
    - angry
    - love
    - neutral
    - boring
    - exciting
    - touching
    - thoughtful
    type: string
    x-enum-comments:
      TitleReactionAngry: "\U0001F620"
      TitleReactionBoring: "\U0001F634"
      TitleReactionDisappointed: "\U0001F61E"
      TitleReactionExciting: "\U0001F929"
      TitleReactionFunny: "\U0001F602"
      TitleReactionInspiring: ✨
      TitleReactionJoyful: "\U0001F60A"
      TitleReactionLove: "\U0001F60D"
      TitleReactionNeutral: "\U0001F610"
      TitleReactionSad: "\U0001F622"
      TitleReactionScary: "\U0001F631"
      TitleReactionSurprised: "\U0001F632"
      TitleReactionThoughtful: "\U0001F914"
      TitleReactionTouching: "\U0001F972"
    x-enum-varnames:
    - TitleReactionSurprised
    - TitleReactionDisappointed
    - TitleReactionSad
    - TitleReactionJoyful
    - TitleReactionInspiring
    - TitleReactionScary
    - TitleReactionFunny
    - TitleReactionAngry
    - TitleReactionLove
    - TitleReactionNeutral
    - TitleReactionBoring
    - TitleReactionExciting
    - TitleReactionTouching
    - TitleReactionThoughtful
  github_com_movie-matcher_backend_internal_models.User:
    properties:
      animeSeriesFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeriesFavorite'
        type: array
      animeSeriesLists:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeriesList'
        type: array
      animeSeriesReviews:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeriesReview'
        type: array
      avatar:
        type: string
      bookFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.BookFavorite'
        type: array
      bookLists:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.BookList'
        type: array
      bookReviews:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.BookReview'
        type: array
      cartoonMovieFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.CartoonMovieFavorite'
        type: array
      cartoonMovieLists:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.CartoonMovieList'
        type: array
      cartoonMovieReviews:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.CartoonMovieReview'
        type: array
      cartoonSeriesFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.CartoonSeriesFavorite'
        type: array
      cartoonSeriesLists:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.CartoonSeriesList'
        type: array
      cartoonSeriesReviews:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.CartoonSeriesReview'
        type: array
      castFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.CastFavorite'
        type: array
      characterFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.CharacterFavorite'
        type: array
      collections:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Collection'
        type: array
      commentBanUntil:
        type: string
      createdAt:
        type: string
      email:
        type: string
      gameFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.GameFavorite'
        type: array
      gameLists:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.GameList'
        type: array
      gameReviews:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.GameReview'
        type: array
      id:
        type: integer
      lightNovelFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LightNovelFavorite'
        type: array
      lightNovelLists:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LightNovelList'
        type: array
      lightNovelReviews:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LightNovelReview'
        type: array
      movieFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.MovieFavorite'
        type: array
      movieLists:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.MovieList'
        type: array
      movieReviews:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.MovieReview'
        type: array
      name:
        type: string
      personFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.PersonFavorite'
        type: array
      role:
        type: string
      settings:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.UserSettings'
      socialLinks:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.SocialLinksMap'
      tvSeriesFavorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeriesFavorite'
        type: array
      tvSeriesLists:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeriesList'
        type: array
      tvSeriesReviews:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeriesReview'
        type: array
      updatedAt:
        type: string
      username:
        description: уникальный для URL, латиница/цифры (nullable для старых записей)
        type: string
    type: object
...
  internal_handlers.ExtendedPaginatedResponse:
    properties:
      data: {}
      filters:
        $ref: '#/definitions/internal_handlers.FilterInfo'
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  internal_handlers.FilterInfo:
    properties:
      ageRatings:
        items:
          type: string
        type: array
      authorIds:
        items:
          type: string
        type: array
      countries:
        items:
          type: string
        type: array
      developerIds:
        items:
          type: string
        type: array
      genreIds:
        items:
          type: string
        type: array
      genreMode:
        type: string
      illustratorIds:
        items:
          type: string
        type: array
      maxRating:
        type: number
      minRating:
        type: number
      order:
        type: string
      platformIds:
        items:
          type: string
        type: array
      publisherIds:
        items:
          type: string
        type: array
      search:
        type: string
      sortBy:
        type: string
      studioIds:
        items:
          type: string
        type: array
      themeIds:
        items:
          type: string
        type: array
      themeMode:
        type: string
      year:
        type: integer
      yearFrom:
        type: integer
      yearTo:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /anime:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: pageSize
        type: integer
      - default: created_at
        description: 'Sort field: created_at, title, rating, release_date, popularity'
        in: query
        name: sortBy
        type: string
      - default: DESC
        description: 'Sort order: ASC, DESC'
        in: query
        name: order
        type: string
      - description: Search in title and description
        in: query
        name: search
        type: string
      - description: Filter by genre ID
        in: query
        name: genreId
        type: integer
      - description: Filter by theme ID
        in: query
        name: themeId
        type: integer
      - description: Filter by release year
        in: query
        name: year
        type: integer
      - description: Filter by release year from
        in: query
        name: yearFrom
        type: integer
      - description: Filter by release year to
        in: query
        name: yearTo
        type: integer
      - description: Minimum rating
        in: query
        name: minRating
        type: number
      - description: Maximum rating
        in: query
        name: maxRating
        type: number
      - description: Filter by age rating
        in: query
        name: ageRating
        type: string
      - description: Filter by studio ID
        in: query
        name: studioId
        type: integer
      - description: Filter by airing status
        in: query
        name: status
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.ExtendedPaginatedResponse'
      summary: Get anime series list
      tags:
      - anime
...
  /anime/{id}:
    get:
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeries'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get anime series by ID
      tags:
      - Anime
...
  /anime/popular:
    get:
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: popular
        description: popular|trending|rating
        in: query
        name: mode
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AnimeSeries'
            type: array
      summary: Get popular anime series
      tags:
      - Anime
  /anime/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.ExtendedPaginatedResponse'
      summary: Search anime series
      tags:
      - Anime
...
swagger: "2.0"